// src/server/fetch-clients/paystack/schemas.ts
export const planIntervalEnum = z.enum([
  "hourly",
  "daily",
  // "weekly",
  "monthly",
  // "quarterly",
  // "biannually",
  "annually",
]);

export const countryEnum = z.enum([
  "ghana",
  "kenya",
  "nigeria",
  "south africa",
]);

export const currencyEnum = z.enum(["XOF", "NGN", "KES", "GHS", "ZAR"]);

export type PlanInterval = z.infer<typeof planIntervalEnum>;

export const subscriptionStatusEnum = z.enum([
  "active",
  "non-renewing",
  "attention",
  "cancelled",
  "completed",
]);

export type SubscriptionStatus = z.infer<typeof subscriptionStatusEnum>;

export const planSchema = z.object({
  id: z.coerce.string(),
  name: z.string(),
  plan_code: z.string(),
  description: z.string().nullable().optional(),
  amount: z.number(),
  interval: planIntervalEnum.describe("Interval for the plan"),
  integration: z.number().optional(),
});